refactor: move FuzzyIndex and Frecency to utils
These are general-purpose data structures, not semi-specific, so relocate
them from semi/ to utils/ alongside the other shared helpers. Frecency is
now generic over its key type (was hardcoded to TokenKey); the elaborator
uses Frecency<TokenKey>.
