+++ crates/reussir-core/src/utils/mod.rs
@@ -1,5 +1,7 @@
 //! Shared utilities for the Reussir core.
 
 pub mod b62;
+pub mod frecency;
+pub mod fuzzy;
 pub mod punycode;
 pub mod string;
